Fixed reference probability for RP55
