+++ examples/reliability_benchmark_table-output.csv
@@ -1,27 +1,27 @@
 ,Exact,FORM,SORM,Monte Carlo,FORM-IS,Subset
-RP8,0.000784,0.000659887791408224,0.0007838036444007566,0.0007000000000000013,0.000709914995334873,0.0009013999999999996
-RP14,0.00752,0.0007003011483118713,0.000699543607859087,0.0007000000000000023,0.0006418773595422634,0.0006956060000000007
-RP22,0.00416,0.006209672105922188,0.004390901764577381,0.005199999999999993,0.004329853055559286,0.004188000000000027
-RP24,0.00286,0.006209245091320783,0.006209245091320783,0.0031000000000000055,0.0026907036789396662,0.0025280000000000033
-RP25,6.14e-06,0.006349315998192334,0.006349315998192334,0.0,3.782390960990971e-05,3.5359999999999966e-05
-RP28,1.46e-07,2.850470029051372e-08,0.0,0.0,1.27100074537508e-07,1.2157641392712904e-07
-RP31,0.00018,0.022750131948179198,0.022750131948179198,0.0027999999999999956,0.003614200521608511,0.003659999999999997
-RP33,0.00257,0.00134989803165187,0.00134989803165187,0.002200000000000001,0.001856178476935175,0.002642703999999992
-RP35,0.00354,0.00134989803165187,0.002134376192618798,0.003399999999999996,0.00242823640033166,0.0035580000000000026
-RP38,0.0081,0.007902212310547153,0.008029356451346456,0.00790000000000002,0.006695970655439684,0.007833999999999985
-RP53,0.0313,0.1180397571257675,0.029861640993453613,0.036072889550018666,0.030546909548680466,0.03296000000000011
-RP55,0.36,0.4992557217282811,7.894596809186417e-06,0.6029411764705884,0.998717236353331,0.573399999999998
-RP54,0.000998,0.0555570374046411,0.0035548110519367576,0.0008999999999999964,0.0008735171714569442,0.0009063999999999984
-RP57,0.0284,0.0486607788566152,0.0486607788566152,0.026238286479250313,0.014977464470527262,0.028219999999999884
-RP75,0.0107,0.0,0.0,0.011518324607329822,0.0,0.010320000000000005
-RP89,0.00543,2.0085940766264634e-09,2.008593952921553e-09,0.006400000000000007,1.4006601417755132e-05,0.0060418919999999975
-RP107,2.92e-07,2.866515718791918e-07,2.866515718791918e-07,0.0,2.633922036909913e-07,3.261463999999994e-07
-RP110,3.19e-05,3.16712418334023e-05,3.16712418334023e-05,0.0,2.845324789518583e-05,3.210000000000002e-05
-RP111,7.65e-07,0.0,0.0,0.0,0.0,9.189132904000018e-07
-RP63,0.000379,0.9999966023268753,0.0,0.0003000000000000029,0.0,0.000419200000000001
-RP91,0.000697,0.0006994296391071298,0.0007011592058218593,0.0003999999999999998,0.0006854523783335723,0.0005914000000000011
-RP60,0.0456,0.044839680802168665,0.04483968318696701,0.03946297803091945,0.04806855515233271,0.048140000000000086
-RP77,2.87e-07,0.0062762084418307535,0.0,0.0,1.7941749848169745e-07,2.779364938703364e-07
-"Four-branch serial system (Waarts, 2000)",0.0021859614549132322,0.0,0.0,0.0019000000000000032,0.0,0.0018379999999999972
-R-S,0.07864960352514257,0.07864960352514253,0.07864960352514253,0.0843263061411549,0.07435618113872226,0.07933999999999995
-Axial stressed beam,0.029198194624830955,0.02998280309217233,0.029332555046970556,0.03202377022119512,0.02988428483842359,0.02746000000000008
+RP8,0.000784,0.000659887791408224,0.0007838036444007566,0.0004,0.000799608268254993,0.0009154000000000007
+RP14,0.00752,0.0007003011483118713,0.000699543607859087,0.0009999999999999996,0.0007735961052119226,0.0007200999999999995
+RP22,0.00416,0.006209672105922188,0.004390901764577381,0.004300000000000002,0.0043137089615818535,0.0038990000000000036
+RP24,0.00286,0.006209245091320783,0.006209245091320783,0.0030000000000000005,0.0028163149039107955,0.0025180000000000007
+RP25,6.14e-06,0.006349315998192334,0.006349315998192334,9.999999999999999e-05,4.476953415752001e-05,3.874121999999999e-05
+RP28,1.46e-07,2.850470029051372e-08,0.0,0.0,1.2636417020624208e-07,1.319033645999998e-07
+RP31,0.00018,0.022750131948179198,0.022750131948179198,0.0033000000000000026,0.0029934546327205092,0.002940000000000004
+RP33,0.00257,0.00134989803165187,0.00134989803165187,0.0032000000000000006,0.002511609133264544,0.002489507999999998
+RP35,0.00354,0.00134989803165187,0.002134376192618798,0.0037000000000000028,0.0023310393448035256,0.003964
+RP38,0.0081,0.007902212310547153,0.008029356451346456,0.006400000000000004,0.008154652828497377,0.008229
+RP53,0.0313,0.1180397571257675,0.029861640993453613,0.029899999999999986,0.03055715867009469,0.0313
+RP55,0.56,0.4992557217282811,7.894596809186417e-06,0.5672000000000001,0.5580281776154343,0.5619000000000004
+RP54,0.000998,0.0555570374046411,0.0035548110519367576,0.0012,0.0009904771883927094,0.0009529999999999993
+RP57,0.0284,0.0486607788566152,0.0486607788566152,0.030299999999999983,0.014645401885757903,0.02864000000000001
+RP75,0.0107,0.0,0.0,0.010200000000000011,0.0,0.009909
+RP89,0.00543,2.0085940766264634e-09,2.008593952921553e-09,0.004200000000000006,3.990897941919434e-05,0.005656999999999999
+RP107,2.92e-07,2.866515718791918e-07,2.866515718791918e-07,0.0,2.9594853800332274e-07,2.7205507259999937e-07
+RP110,3.19e-05,3.16712418334023e-05,3.16712418334023e-05,0.0,3.270338645572822e-05,2.1371282964359954e-06
+RP111,7.65e-07,0.0,0.0,0.0,0.0,9.275109605999989e-07
+RP63,0.000379,0.9999966023268753,0.0,0.0003,0.0,0.0003734999999999997
+RP91,0.000697,0.0006994296391071298,0.0007011592058218593,0.0007999999999999999,0.0006726034960586403,0.000797199999999999
+RP60,0.0456,0.044839680802168665,0.04483968318696701,0.04879999999999999,0.04514307345264836,0.04478000000000004
+RP77,2.87e-07,0.0062762084418307535,0.0,0.0,2.2759944193758046e-07,1.871126999999999e-07
+Four-branch serial system,0.0021859614549132322,0.0,0.0,0.002300000000000001,0.0,0.0020049999999999994
+R-S,0.07864960352514257,0.07864960352514253,0.07864960352514253,0.07509999999999999,0.07964097684398252,0.07841999999999999
+Axial stressed beam,0.029198194624830955,0.02998280309217233,0.029332555046970556,0.03089999999999998,0.03031439159547681,0.029969999999999993
